Module teaclave_config::build

Structs

BuildConfig

Inbounds

Constants

ACCESS_CONTROL_INBOUND_SERVICES
Array of predefined inbound services, usually used for validate incoming connections via mutual attestation.

AS_ROOT_CA_CERT
CA certification of Attestation Service in binary (DER format).

ATTESTATION_VALIDITY_SECS
The valid duration of one attestation report in seconds.

AUDITOR_PUBLIC_KEYS
Array of auditor's public keys in binary (DER format), usually used to verify signatures of enaclave_info.toml.

AUTHENTICATION_INBOUND_SERVICES
Array of predefined inbound services, usually used for validate incoming connections via mutual attestation.

BUILD_CONFIG

MANAGEMENT_INBOUND_SERVICES
Array of predefined inbound services, usually used for validate incoming connections via mutual attestation.

SCHEDULER_INBOUND_SERVICES
Array of predefined inbound services, usually used for validate incoming connections via mutual attestation.

STORAGE_INBOUND_SERVICES
Array of predefined inbound services, usually used for validate incoming connections via mutual attestation.